make address and address modal components
