+++ projects/client/src/pages/user/PrescriptionPage.jsx
@@ -8,13 +8,15 @@ import bni from '../../assets/Bank BNI Logo (PNG-1080p) - FileVector69.png';
 import bca from '../../assets/Bank BCA Logo (PNG-1080p) - FileVector69.png';
 import bri from '../../assets/bri.png';
 import LoadingComponent from "../../components/Loading";
+import NewAddressComponent from "../../components/NewAddressModalComp";
+import EditAddressComponent from "../../components/EditAddressModalComp";
 
 const Prescription = (props) => {
 
     const [showPic, setShowPic] = React.useState('');
     const [prescriptionPic, setPrescriptionPic] = React.useState('');
     const [loadPic, setLoadPic] = React.useState(false);
-    
+
     const [loading, setLoading] = React.useState(false);
 
     const [allAddress, setAllAddress] = React.useState([]);
@@ -26,9 +28,6 @@ const Prescription = (props) => {
     const [paymentMethod, setPaymentMethod] = React.useState('');
 
     // RAJAONGKIR
-    const [dataProvince, setDataProvince] = React.useState([]);
-    const [dataCity, setDataCity] = React.useState([]);
-    const [filterCity, setFilterCity] = React.useState([]);
     const [courier, setCourier] = React.useState('');
     const [delivery, setDelivery] = React.useState([]);
     const [selectedDelivery, setSelectedDelivery] = React.useState('');
@@ -38,47 +37,14 @@ const Prescription = (props) => {
     const [origin, setOrigin] = React.useState(153); //kode city_id Jakarta Selatan
     const [totalDelivery, setTotalDelivery] = React.useState(0);
 
-    // INPUT NEW ADDRESS
-    const [countFullAddress, setCountFullAddress] = React.useState(0);
-    const [selectedProvinceID, setSelectedProvinceID] = React.useState(0);
-    const [selectedCityID, setSelectedCityID] = React.useState(0);
-    const [inputFullAddress, setInputFullAddress] = React.useState('');
-    const [inputDistrict, setInputDistrict] = React.useState('');
-    const [inputPostalCode, setInputPostalCode] = React.useState('');
-    const [checkAddress, setCheckAddress] = React.useState('');
-    const [checkProvince, setCheckProvince] = React.useState('');
-    const [checkCity, setCheckCity] = React.useState('');
-    const [checkDistrict, setCheckDistrict] = React.useState('');
-    const [checkPostal, setCheckPostal] = React.useState('');
-
     // INPUT EDIT ADDRESS
     const [selectedEdit, setSelectedEdit] = React.useState({});
-    const [countEditFullAddress, setCountEditFullAddress] = React.useState(0);
-    const [selectedEditProvinceID, setSelectedEditProvinceID] = React.useState(0);
-    const [selectedEditCityID, setSelectedEditCityID] = React.useState(0);
-    const [inputEditFullAddress, setInputEditFullAddress] = React.useState('');
-    const [inputEditDistrict, setInputEditDistrict] = React.useState('');
-    const [inputEditPostalCode, setInputEditPostalCode] = React.useState('');
-
-    const [showEditFullAddress, setShowEditFullAddress] = React.useState('');
-    const [showEditProvince, setShowEditProvince] = React.useState('');
-    const [showEditCity, setShowEditCity] = React.useState('');
-    const [showEditDistrict, setShowEditDistrict] = React.useState('');
-    const [showEditPostal, setShowEditPostal] = React.useState('');
-
-    const [checkEditAddress, setCheckEditAddress] = React.useState('');
-    const [checkEditProvince, setCheckEditProvince] = React.useState('');
-    const [checkEditCity, setCheckEditCity] = React.useState('');
-    const [checkEditDistrict, setCheckEditDistrict] = React.useState('');
-    const [checkEditPostal, setCheckEditPostal] = React.useState('');
 
     const fileRef = React.useRef();
     const navigate = useNavigate();
 
     React.useEffect(() => {
         getAddress();
-        getDataProvince();
-        getDataCity();
     }, []);
 
     const getAddress = async () => {
@@ -114,7 +80,7 @@ const Prescription = (props) => {
     const onSelectAddress = async (id) => {
         try {
             let userToken = localStorage.getItem('medcarelog');
-            let select = await axios.patch(API_URL + '/api/user/updateaddress', { selected: 'true', idaddress: id }, {
+            let select = await axios.patch(API_URL + '/api/address/update', { selected: 'true', idaddress: id }, {
                 headers: {
                     'Authorization': `Bearer ${userToken}`
                 }
@@ -181,203 +147,6 @@ const Prescription = (props) => {
                     You don't have any address. Please click select address and click add new address.
                 </div>
             )
-        }
-    };
-
-    const getDataProvince = async () => {
-        try {
-            let province = await axios.get(API_URL + '/api/rajaongkir/province');
-
-            //console.log(province.data)
-            setDataProvince(province.data)
-        } catch (error) {
-            console.log(error)
-        }
-    };
-
-    const getDataCity = async () => {
-        try {
-            let city = await axios.get(API_URL + '/api/rajaongkir/city');
-
-            //console.log(city.data)
-            setDataCity(city.data)
-        } catch (error) {
-            console.log(error)
-        }
-    };
-
-    const handleFilterCity = (id) => {
-        //console.log(id)
-        let temp = [];
-
-        dataCity.forEach((val, idx) => {
-            if (val.province_id === id) {
-                temp.push(val);
-            }
-        })
-        //console.log(temp);
-        setFilterCity(temp);
-    };
-
-    const onSaveNewAddress = async () => {
-        try {
-            let userToken = localStorage.getItem('medcarelog');
-            // console.log("full", inputFullAddress); string
-            // console.log("district", inputDistrict); string
-            // console.log("postal", inputPostalCode); number
-            // console.log("cityID", selectedCityID); number
-            // console.log("provinceID", selectedProvinceID); number
-
-            let searchData = dataProvince.find(val => val.province_id === selectedProvinceID);
-            let searchCity = filterCity.find(val => val.city_id === selectedCityID);
-
-            // console.log("kota", searchCity.type + ' ' + searchCity.city_name);
-
-            if (inputFullAddress === '' || inputDistrict === '' || searchData === undefined || searchCity === undefined || inputPostalCode === '') {
-                if (inputFullAddress === '') {
-                    setCheckAddress('show');
-                }
-                if (inputDistrict === '') {
-                    setCheckDistrict('show');
-                }
-                if (searchData === undefined) {
-                    setCheckProvince('show');
-                }
-                if (searchCity === undefined) {
-                    setCheckCity('show');
-                }
-                if (inputPostalCode === '') {
-                    setCheckPostal('show');
-                }
-            } else {
-                let data = {
-                    full_address: inputFullAddress,
-                    district: inputDistrict,
-                    city: searchCity.type + ' ' + searchCity.city_name,
-                    city_id: selectedCityID,
-                    province: searchData.province,
-                    province_id: selectedProvinceID,
-                    postal_code: inputPostalCode
-                }
-
-                // pakai authorization
-                let add = await axios.post(API_URL + '/api/address/add', { data }, {
-                    headers: {
-                        'Authorization': `Bearer ${userToken}`
-                    }
-                })
-
-                if (add.data.success) {
-                    getAddress();
-                    toast.success('Address Added Success!', {
-                        theme: 'colored',
-                        position: "top-center",
-                        autoClose: 2000,
-                        hideProgressBar: false,
-                        closeOnClick: true,
-                        pauseOnHover: true,
-                        draggable: true,
-                        progress: undefined,
-                    });
-                    setInputFullAddress('');
-                    setInputDistrict('');
-                    setInputPostalCode('');
-                    setSelectedProvinceID(0);
-                    setSelectedCityID(0);
-                    setCountFullAddress(0);
-                    setCheckAddress('');
-                    setCheckCity('');
-                    setCheckDistrict('');
-                    setCheckPostal('');
-                    setCheckProvince('');
-                    setShowNewAddressModal('');
-                    setShowAddressModal('show');
-                }
-
-                console.log(data)
-            }
-
-        } catch (error) {
-            console.log(error)
-        }
-    };
-
-    const onSaveEditAddress = async () => {
-        try {
-            let userToken = localStorage.getItem('medcarelog');
-            // console.log("full", inputEditFullAddress);
-            // console.log("district", inputEditDistrict);
-            // console.log("postal", inputEditPostalCode);
-            // console.log("cityID", selectedEditCityID);
-            // console.log("provinceID", selectedEditProvinceID);
-            // console.log(selectedEdit)
-
-            let searchData = dataProvince.find(val => val.province_id === selectedEditProvinceID);
-            let searchCity = filterCity.find(val => val.city_id === selectedEditCityID);
-
-            let tempCity = '';
-            if (searchCity === undefined) {
-                tempCity = '';
-            } else {
-                tempCity = searchCity.type + ' ' + searchCity.city_name;
-            }
-
-            let tempProvince = '';
-            if (searchData === undefined) {
-                tempProvince = '';
-            } else {
-                tempProvince = searchData.province;
-            }
-
-            let dataEdit = {
-                full_address: inputEditFullAddress,
-                district: inputEditDistrict,
-                city: tempCity,
-                city_id: selectedEditCityID,
-                province: tempProvince,
-                province_id: selectedEditProvinceID,
-                postal_code: inputEditPostalCode
-            }
-
-            // // pakai authorization
-            if (tempProvince !== '' && tempCity === '') {
-                setCheckEditCity('show');
-            } else {
-                let edit = await axios.patch(API_URL + '/api/address/update', { dataEdit, idaddress: selectedEdit.idaddress }, {
-                    headers: {
-                        'Authorization': `Bearer ${userToken}`
-                    }
-                })
-
-                if (edit.data.success) {
-                    getAddress();
-                    toast.success('Address Edit Success!', {
-                        theme: 'colored',
-                        position: "top-center",
-                        autoClose: 2000,
-                        hideProgressBar: false,
-                        closeOnClick: true,
-                        pauseOnHover: true,
-                        draggable: true,
-                        progress: undefined,
-                    });
-                    setShowAddressModal('show');
-                    setShowEditAddressModal('');
-                    setShowEditFullAddress('');
-                    setShowEditCity('');
-                    setShowEditDistrict('');
-                    setShowEditProvince('');
-                    setShowEditPostal('');
-                    setInputEditFullAddress('');
-                    setInputEditDistrict('');
-                    setInputEditPostalCode('');
-                    setSelectedEditCityID(0);
-                    setSelectedEditProvinceID(0);
-                }
-            }
-            console.log(dataEdit)
-        } catch (error) {
-            console.log(error)
         }
     };
 
@@ -472,15 +241,15 @@ const Prescription = (props) => {
             setLoading(true);
             setShowPaymentModal('');
             let res = await axios.post(API_URL + '/api/transaction/addprescription', formPay, {
-                headers : {
-                    'Authorization' : `Bearer ${userToken}`
+                headers: {
+                    'Authorization': `Bearer ${userToken}`
                 }
             })
 
-            if (res.data.success){
+            if (res.data.success) {
                 setTimeout(() => {
                     setLoading(false);
-                    navigate('/prescription/success', {replace: true})
+                    navigate('/prescription/success', { replace: true })
                 }, 1000)
             }
         } catch (error) {
@@ -529,286 +298,14 @@ const Prescription = (props) => {
                         {/* MODAL ALAMAT BARU */}
                         {
                             showNewAddressModal === 'show' ?
-                                <div tabIndex={-1} className="overflow-y-auto overflow-x-hidden backdrop-blur-sm fixed right-0 left-0 top-0 flex justify-center items-center z-50 md:inset-0 h-modal md:h-full">
-                                    <div className="relative p-4 w-1/2 h-full md:h-auto">
-                                        <div className="relative border-2 bg-white rounded-lg shadow border-main-500">
-                                            <div className="p-6 text-center">
-                                                <div>
-                                                    <p className='text-2xl font-bold text-main-500'>Add New Address</p>
-                                                </div>
-                                                <div className='border my-4'>
-                                                    <div className='border flex flex-col items-start px-3 mb-2'>
-                                                        <p>Full Address :</p>
-                                                        <textarea maxLength={200} type='text'
-                                                            className={checkAddress === 'show' ? 'border border-red-600 w-full rounded-lg px-3 mt-2' : 'border border-main-600 w-full rounded-lg px-3 mt-2 focus:ring-2 focus:ring-main-500'}
-                                                            placeholder='Full Address' onChange={(e) => { setInputFullAddress(e.target.value); setCountFullAddress(e.target.value.length); if (e.target.value.length > 0) { setCheckAddress('') } }} value={inputFullAddress} />
-                                                        <div className='flex justify-end w-full'>
-                                                            {countFullAddress} / 200
-                                                        </div>
-                                                    </div>
-                                                    <div className='border flex flex-col items-start px-3 mb-2'>
-                                                        <p>Province (Provinsi) :</p>
-                                                        <select type='text' onChange={(e) => { handleFilterCity(e.target.value); setSelectedProvinceID(e.target.value); if (e.target.value > 0) { setCheckProvince('') } }}
-                                                            className={checkProvince === 'show' ? 'border border-red-600 w-full rounded-lg px-3 h-10 mt-2' :
-                                                                'border border-main-600 w-full rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}>
-                                                            <option value='' className='w-1/2'>Select Province</option>
-                                                            {
-                                                                dataProvince.map((val, idx) => {
-                                                                    return (
-                                                                        <option value={val.province_id} key={val.province_id}>{val.province}</option>
-                                                                    )
-                                                                })
-                                                            }
-                                                        </select>
-                                                    </div>
-                                                    <div className='border flex flex-col items-start px-3 mb-2'>
-                                                        <p>City (Kota) :</p>
-                                                        <select type='text' onChange={(e) => { setSelectedCityID(e.target.value); if (e.target.value > 0) { setCheckCity('') } }}
-                                                            className={checkCity === 'show' ? 'border border-red-600 w-full rounded-lg px-3 h-10 mt-2' :
-                                                                'border border-main-600 w-full rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}>
-                                                            <option value='' className='w-1/2'>Select City</option>
-                                                            {
-                                                                filterCity.map((val, idx) => {
-                                                                    return (
-                                                                        <option value={val.city_id} key={val.city_id}>{val.type} {val.city_name}</option>
-                                                                    )
-                                                                })
-                                                            }
-                                                        </select>
-                                                    </div>
-                                                    <div className='border flex flex-col items-start px-3 mb-2'>
-                                                        <p>District (Kecamatan/Kabupaten) :</p>
-                                                        <input type='text'
-                                                            className={checkDistrict === 'show' ? 'border border-red-600 w-full rounded-lg px-3 h-10 mt-2' :
-                                                                'border border-main-600 w-full rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}
-                                                            placeholder='District' onChange={(e) => { setInputDistrict(e.target.value); if (e.target.value.length > 0) { setCheckDistrict('') } }}
-                                                            value={inputDistrict} />
-                                                    </div>
-                                                    <div className='border flex flex-col items-start px-3 mb-2'>
-                                                        <p>Postal Code (Kode Pos) :</p>
-                                                        <input type='number'
-                                                            className={checkPostal === 'show' ? 'border border-red-600 w-full rounded-lg px-3 h-10 mt-2' :
-                                                                'border border-main-600 w-full rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}
-                                                            placeholder='Postal Code' onChange={(e) => { setInputPostalCode(e.target.value); if (e.target.value > 0) { setCheckPostal('') } }}
-                                                            value={inputPostalCode} />
-                                                    </div>
-                                                </div>
-                                                <button type="button" className="mr-1 text-white bg-main-500 focus:ring-4 focus:outline-none focus:ring-gray-200 rounded-lg border border-gray-200 text-sm font-medium px-10 py-2.5 focus:z-10 "
-                                                    onClick={() => {
-                                                        onSaveNewAddress();
-                                                    }}>Save</button>
-                                                <button type="button" className="ml-1 text-black bg-white focus:ring-4 focus:outline-none focus:ring-gray-200 rounded-lg border border-gray-200 text-sm font-medium px-10 py-2.5 focus:z-10 "
-                                                    onClick={() => {
-                                                        setInputFullAddress('');
-                                                        setInputDistrict('');
-                                                        setInputPostalCode('');
-                                                        setSelectedProvinceID(0);
-                                                        setSelectedCityID(0);
-                                                        setCountFullAddress(0);
-                                                        setCheckAddress('');
-                                                        setCheckCity('');
-                                                        setCheckDistrict('');
-                                                        setCheckPostal('');
-                                                        setCheckProvince('');
-                                                        setShowAddressModal('show');
-                                                        setShowNewAddressModal('')
-                                                    }}
-                                                >Cancel</button>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
+                                <NewAddressComponent showModal={setShowNewAddressModal} setAddress={getAddress()} />
                                 :
                                 ""
                         }
                         {/* MODAL EDIT ALAMAT */}
                         {
                             showEditAddressModal === 'show' ?
-                                <div tabIndex={-1} className="overflow-y-auto overflow-x-hidden backdrop-blur-sm fixed right-0 left-0 top-0 flex justify-center items-center z-50 md:inset-0 h-modal md:h-full">
-                                    <div className="relative p-4 w-1/2 h-full md:h-auto">
-                                        <div className="relative border-2 bg-white rounded-lg shadow border-main-500">
-                                            <div className="p-6 text-center">
-                                                <div>
-                                                    <p className='text-2xl font-bold text-main-500'>Edit Address</p>
-                                                </div>
-                                                <div className='my-4'>
-                                                    <div className='flex flex-col items-start px-3 mb-2'>
-                                                        {
-                                                            showEditFullAddress === 'show' ?
-                                                                <>
-                                                                    <p>Full Address :</p>
-                                                                    <div className='w-full flex'>
-                                                                        <div className='w-11/12'>
-                                                                            <textarea maxLength={200} type='text'
-                                                                                className={checkEditAddress === 'show' ? 'border border-red-600 w-full rounded-lg px-3 mt-2' : 'border border-main-600 w-full rounded-lg px-3 mt-2 focus:ring-2 focus:ring-main-500'}
-                                                                                placeholder={selectedEdit.full_address} onChange={(e) => { setInputEditFullAddress(e.target.value); setCountEditFullAddress(e.target.value.length); if (e.target.value.length > 0) { setCheckEditAddress('') } }}
-                                                                                value={inputEditFullAddress} />
-                                                                            <div className='flex justify-end w-full'>
-                                                                                {countEditFullAddress} / 200
-                                                                            </div>
-                                                                        </div>
-                                                                        <div className='w-1/12 flex items-center justify-center'>
-                                                                            <button type='button' className='text-red-600 hover:underline ' onClick={() => { setShowEditFullAddress(''); setInputEditFullAddress('') }}>Cancel</button>
-                                                                        </div>
-                                                                    </div>
-                                                                </>
-                                                                :
-                                                                <div className='flex w-full py-3'>
-                                                                    <div className='w-11/12 flex text-start'>
-                                                                        <p className='w-1/6 font-bold'>Full Address</p>
-                                                                        <p>: {selectedEdit.full_address}</p>
-                                                                    </div>
-                                                                    <div className='w-1/12 flex items-center justify-center'>
-                                                                        <button type='button' className='text-main-600 hover:underline ' onClick={() => setShowEditFullAddress('show')}>Edit</button>
-                                                                    </div>
-                                                                </div>
-                                                        }
-                                                    </div>
-                                                    <div className='flex flex-col items-start px-3 mb-2'>
-                                                        {
-                                                            showEditProvince === 'show' ?
-                                                                <>
-                                                                    <p>Province (Provinsi) :</p>
-                                                                    <div className='w-full flex'>
-                                                                        <select type='text' onChange={(e) => { handleFilterCity(e.target.value); setSelectedEditProvinceID(e.target.value); if (e.target.value > 0) { setCheckEditProvince('') } }}
-                                                                            className={checkEditProvince === 'show' ? 'w-11/12 border border-red-600 rounded-lg px-3 h-10 mt-2' :
-                                                                                'border border-main-600 w-11/12 rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}>
-                                                                            <option value='' className='w-1/2'>Select Province</option>
-                                                                            {
-                                                                                dataProvince.map((val, idx) => {
-                                                                                    return (
-                                                                                        <option value={val.province_id} key={val.province_id}>{val.province}</option>
-                                                                                    )
-                                                                                })
-                                                                            }
-                                                                        </select>
-                                                                        <div className='w-1/12 flex items-center justify-center'>
-                                                                            <button type='button' className='text-red-600 hover:underline'
-                                                                                onClick={() => { setShowEditProvince(''); setSelectedEditProvinceID(0); setSelectedEditCityID(0) }}>Cancel</button>
-                                                                        </div>
-                                                                    </div>
-                                                                </>
-                                                                :
-                                                                <div className='flex w-full py-3'>
-                                                                    <div className='w-11/12 flex text-start'>
-                                                                        <p className='w-1/6 font-bold'>Province</p>
-                                                                        <p>: {selectedEdit.province}</p>
-                                                                    </div>
-                                                                    <div className='w-1/12 flex items-center justify-center'>
-                                                                        <button type='button' className='text-main-600 hover:underline ' onClick={() => setShowEditProvince('show')}>Edit</button>
-                                                                    </div>
-                                                                </div>
-                                                        }
-                                                    </div>
-                                                    <div className='flex flex-col items-start px-3 mb-2'>
-                                                        {
-                                                            showEditProvince === 'show' ?
-                                                                <>
-                                                                    <p>City (Kota) :</p>
-                                                                    <div className='w-full flex'>
-                                                                        <select type='text' onChange={(e) => { setSelectedEditCityID(e.target.value); if (e.target.value > 0) { setCheckEditCity('') } }}
-                                                                            className={checkEditCity === 'show' ? 'border border-red-600 w-11/12 rounded-lg px-3 h-10 mt-2' :
-                                                                                'border border-main-600 w-11/12 rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}>
-                                                                            <option value='' className='w-1/2'>Select City</option>
-                                                                            {
-                                                                                filterCity.map((val, idx) => {
-                                                                                    return (
-                                                                                        <option value={val.city_id} key={val.city_id}>{val.type} {val.city_name}</option>
-                                                                                    )
-                                                                                })
-                                                                            }
-                                                                        </select>
-                                                                    </div>
-                                                                </>
-                                                                :
-                                                                <div className='flex w-full py-3'>
-                                                                    <div className='w-11/12 flex text-start'>
-                                                                        <p className='w-1/6 font-bold'>City</p>
-                                                                        <p>: {selectedEdit.city}</p>
-                                                                    </div>
-                                                                </div>
-                                                        }
-                                                    </div>
-                                                    <div className='flex flex-col items-start px-3 mb-2'>
-                                                        {
-                                                            showEditDistrict === 'show' ?
-                                                                <>
-                                                                    <p>District (Kecamatan/Kabupaten) :</p>
-                                                                    <div className='w-full flex'>
-                                                                        <input type='text'
-                                                                            className={checkEditDistrict === 'show' ? 'border border-red-600 w-11/12 rounded-lg px-3 h-10 mt-2' :
-                                                                                'border border-main-600 w-11/12 rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}
-                                                                            placeholder={selectedEdit.district} onChange={(e) => { setInputEditDistrict(e.target.value); if (e.target.value.length > 0) { setCheckEditDistrict('') } }}
-                                                                            value={inputEditDistrict} />
-                                                                        <div className='w-1/12 flex items-center justify-center'>
-                                                                            <button type='button' className='text-red-600 hover:underline ' onClick={() => { setShowEditDistrict(''); setInputEditDistrict('') }}>Cancel</button>
-                                                                        </div>
-                                                                    </div>
-                                                                </>
-                                                                :
-                                                                <div className='flex w-full py-3'>
-                                                                    <div className='w-11/12 flex text-start'>
-                                                                        <p className='w-1/6 font-bold'>District</p>
-                                                                        <p>: {selectedEdit.district}</p>
-                                                                    </div>
-                                                                    <div className='w-1/12 flex items-center justify-center'>
-                                                                        <button type='button' className='text-main-600 hover:underline ' onClick={() => setShowEditDistrict('show')}>Edit</button>
-                                                                    </div>
-                                                                </div>
-                                                        }
-                                                    </div>
-                                                    <div className='flex flex-col items-start  px-3 mb-2'>
-                                                        {
-                                                            showEditPostal === 'show' ?
-                                                                <>
-                                                                    <p>Postal Code (Kode Pos) :</p>
-                                                                    <div className='w-full flex'>
-                                                                        <input type='number'
-                                                                            className={checkEditPostal === 'show' ? 'border border-red-600 w-11/12 rounded-lg px-3 h-10 mt-2' :
-                                                                                'border border-main-600 w-11/12 rounded-lg px-3 h-10 mt-2 focus:ring-2 focus:ring-main-500'}
-                                                                            placeholder={selectedEdit.postal_code} onChange={(e) => { setInputEditPostalCode(e.target.value); if (e.target.value > 0) { setCheckEditPostal('') } }}
-                                                                            value={inputEditPostalCode} />
-                                                                        <div className='w-1/12 flex items-center justify-center'>
-                                                                            <button type='button' className='text-red-600 hover:underline ' onClick={() => { setShowEditPostal(''); setInputEditPostalCode('') }}>Cancel</button>
-                                                                        </div>
-                                                                    </div>
-                                                                </>
-                                                                :
-                                                                <div className='flex w-full py-3'>
-                                                                    <div className='w-11/12 flex text-start'>
-                                                                        <p className='w-1/6 font-bold'>Postal Code</p>
-                                                                        <p>: {selectedEdit.postal_code}</p>
-                                                                    </div>
-                                                                    <div className='w-1/12 flex items-center justify-center'>
-                                                                        <button type='button' className='text-main-600 hover:underline ' onClick={() => setShowEditPostal('show')}>Edit</button>
-                                                                    </div>
-                                                                </div>
-                                                        }
-                                                    </div>
-                                                </div>
-                                                <button type="button" className="mr-1 text-white bg-main-500 focus:ring-4 focus:outline-none focus:ring-gray-200 rounded-lg border border-gray-200 text-sm font-medium px-10 py-2.5 focus:z-10 disabled:bg-opacity-50"
-                                                    onClick={() => onSaveEditAddress()} disabled={showEditFullAddress === 'show' || showEditDistrict === 'show' || showEditPostal === 'show' || showEditProvince === 'show' ? false : true}
-                                                >Save</button>
-                                                <button type="button" className="ml-1 text-black bg-white focus:ring-4 focus:outline-none focus:ring-gray-200 rounded-lg border border-gray-200 text-sm font-medium px-10 py-2.5 focus:z-10 "
-                                                    onClick={() => {
-                                                        setShowAddressModal('show');
-                                                        setShowEditAddressModal('');
-                                                        setShowEditFullAddress('');
-                                                        setShowEditCity('');
-                                                        setShowEditDistrict('');
-                                                        setShowEditProvince('');
-                                                        setShowEditPostal('');
-                                                        setInputEditFullAddress('');
-                                                        setInputEditDistrict('');
-                                                        setInputEditPostalCode('');
-                                                        setSelectedEditCityID(0);
-                                                        setSelectedEditProvinceID(0);
-                                                    }}
-                                                >Cancel</button>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
+                                <EditAddressComponent selected={selectedEdit} showModal={setShowEditAddressModal} setAddress={getAddress()} />
                                 :
                                 ""
                         }
@@ -1040,7 +537,7 @@ const Prescription = (props) => {
                 </div>
             </div>
             <ToastContainer />
-            <LoadingComponent loading={loading}/>
+            <LoadingComponent loading={loading} />
         </div>
     )
